Revert "PLAT-1797"
